changes to models
