+++ use_pySODM/models.py
@@ -14,7 +14,7 @@ from pySODM.models.base import ODE, JumpProcess
 
 class spatial_ODE_SIR(ODE):
     """
-    SIR model with a spatial stratification
+    SIR model with age and spatial stratification
     """
     
     states = ['S','I','R']
@@ -24,24 +24,46 @@ class spatial_ODE_SIR(ODE):
     @staticmethod
     def integrate(t, S, I, R, beta, gamma, f_v, N, M):
 
-        # compute total population 
-        T = S + I + R
+        # compute contact tensor with different home vs. visited contacts
+        C =  ((1 - f_v) * np.einsum('ab,cd->abcd', N, np.eye(M.shape[0])) + f_v * np.einsum('ab,cd->abcd', N, M))
 
-        # compute visiting populations
-        T_v = matmul_2D_3D_matrix(T, M) # M can  be of size (n_loc, n_loc) or (n_loc, n_loc, n_age), representing a different OD matrix in every age group
-        S_v = matmul_2D_3D_matrix(S, M)
-        I_v = matmul_2D_3D_matrix(I, M)
+        # compute force of infection
+        l = beta * np.einsum ('abcd,bd->ac', C, I/(S+I+R))
 
-        # compute number of new infections on home patch and visited patch
-        dI_h = beta * S * np.transpose(matmul_2D_3D_matrix(np.transpose(I/T), (1-f_v)*N)) # N can  be of size (n_age, n_age) or (n_age, n_age, n_loc), representing a different contact matrix in every spatial patch 
-        dI_v = beta * S_v * np.transpose(matmul_2D_3D_matrix(np.transpose(I_v/T_v), f_v*N))
+        # calculate differentials
+        dS = - l * S
+        dI = l * S - 1/gamma*I
+        dR = 1/gamma*I
 
-        # distribute the number of new infections on visited patch to the home patch 
-        dI_v = S * np.transpose(np.atleast_2d(M) @ np.transpose(dI_v/S_v))
+        return dS, dI, dR
 
-        # Calculate differentials
-        dS = - (dI_h + dI_v)
-        dI = (dI_h + dI_v) - 1/gamma*I
+# tensorflow variant --> can utilise the GPU but setup for macOS is a pain
+import tensorflow as tf
+class spatial_ODE_SIR_tf(ODE):
+    """
+    SIR model with age and spatial stratification
+    """
+    
+    states = ['S','I','R']
+    parameters = ['beta','gamma', 'f_v', 'N', 'M']
+    dimensions = ['age', 'location']
+
+    @staticmethod
+    def integrate(t, S, I, R, beta, gamma, f_v, N, M):
+
+        # conversion to tensors (ideally N, M outside script)
+        N = tf.convert_to_tensor(N, dtype=float)
+        M = tf.convert_to_tensor(M, dtype=float)
+        I_rel = tf.convert_to_tensor(I/(S+I+R), dtype=float)
+
+        # compute contact tensor with different home vs. visited contacts
+        C =  ((1 - f_v) * tf.einsum('ab,cd->abcd', N, tf.eye(M.shape[0])) + f_v * tf.einsum('ab,cd->abcd', N, M))
+        # compute force of infection
+        l = beta * tf.einsum ('abcd,bd->ac', C, I_rel)
+
+        # calculate differentials
+        dS = - l * S
+        dI = l * S - 1/gamma*I
         dR = 1/gamma*I
 
         return dS, dI, dR
@@ -50,84 +72,36 @@ class spatial_ODE_SIR(ODE):
 ### Stochastic ###
 ###################
 
-
 class spatial_TL_SIR(JumpProcess):
     """
-    SIR stochastic model with a spatial stratification
+    Stochastic SIR model with age and spatial stratification
     """
-    states = ['S', 'S_v','I','R']
+    states = ['S', 'I','R']
     parameters = ['beta','gamma', 'f_v', 'N', 'M']
     dimensions = ['age', 'location']
 
 
     @staticmethod
-    def compute_rates(t, S, S_v, I, R, beta, gamma, f_v, N, M):
+    def compute_rates(t, S, I, R, beta, gamma, f_v, N, M):
 
-        # calculate total population 
-        T = S + I + R
+        # compute contact tensor with different home vs. visited contacts
+        C =  ((1 - f_v) * np.einsum('ab,cd->abcd', N, np.eye(M.shape[0])) + f_v * np.einsum('ab,cd->abcd', N, M))
 
-        # compute visiting populations
-        T_v = matmul_2D_3D_matrix(T, M) # M can  be of size (n_loc, n_loc) or (n_loc, n_loc, n_age), representing a different OD matrix in every age group
-        I_v = matmul_2D_3D_matrix(I, M)
-
-        # create a size dummy 
-        G = S.shape[0] # age stratification
-        H = S.shape[1] # spatial stratification
-        size_dummy = np.ones([G,H], np.float64)
+        # compute force of infection
+        l = beta * np.einsum('abcd,bd->ac', C, I/(S+I+R))
 
         rates = {
-
-            'S': [beta * np.transpose(matmul_2D_3D_matrix(np.transpose(I/T), (1-f_v)*N))], # 
-            'S_v': [beta * np.transpose(matmul_2D_3D_matrix(np.transpose(I_v/T_v), f_v*N))],
-            'I': [size_dummy*(1/gamma)], # 
-
+            'S': [l], 
+            'I': [np.ones(S.shape, np.float64)*(1/gamma)], 
             }
         
         return rates
 
     @ staticmethod
-    def apply_transitionings(t, tau, transitionings, S, S_v, I, R, 
-                             beta, f_v, gamma, 
-                             N, M):
+    def apply_transitionings(t, tau, transitionings, S, I, R, beta, f_v, gamma, N, M):
         
-        # distribute the number of new infections on visited patch to the home patch 
-        S_v_to_home = S * np.transpose(np.atleast_2d(M) @ np.transpose(transitionings['S_v'][0]/S_v))
-        # the resulting matrix is an N x M matrix with each element of row n, column m representing the total number of people infected from work that need to be returned to age-group n, location m. 
-        
-        
-        ###################################################
-        ##### CREATE THE NEW VALUES FOR S, S_v, I AND R
-        ###################################################
-
-        S_new = S - transitionings['S'][0] - S_v_to_home[0]
-        S_new = np.where(S_new < 0, 0, S_new)
-
-        S_v_new =  matmul_2D_3D_matrix(S_new, M)
-        I_new = I + transitionings['S'][0] + S_v_to_home[0] - transitionings['I'][0]
+        S_new = S - transitionings['S'][0]
+        I_new = I + transitionings['S'][0] - transitionings['I'][0]
         R_new = R + transitionings['I'][0]
         
-        return(S_new,S_v_new, I_new, R_new)
-    
-# helper function
-def matmul_2D_3D_matrix(X, W):
-    """
-    Computes the product of a 2D matrix (size n x m) and a 3D matrix (size m x m x n) as an n-dimensional stack of (1xm) and (m,m) products.
-
-    input
-    =====
-    X: np.ndarray
-        Matrix of size (n,m).
-    W : np.ndarray
-        2D or 3D matrix:
-        - If 2D: Shape (m, m). Expanded to size (m, m, n).
-        - If 3D: Shape (m, m, n).
-          Represents n stacked (m x m) matrices.
-
-    output
-    ======
-    X_out : np.ndarray
-        Matrix product of size (n, m). 
-        Element-wise equivalent operation: O_{ij} = \sum_{l} [ s_{il} * w_{lji} ]
-    """
-    W = np.atleast_3d(W)
-    return np.einsum('ik,kji->ij', X, np.broadcast_to(W, (W.shape[0], W.shape[0], X.shape[0])))
+        return(S_new, I_new, R_new)
